feat: multithreading (#32)
* feat: wip multithreading

* fixup! feat: wip multithreading

* feat: rework with batched lock free queue

* feat: expandable queue implementation

* feat: tune thread count for parallel matcher

* feat: re-order benches

* feat: rework benchmarks

* feat: slice for fixed queue, parallel sorting, reduce match size

* feat: misc

* feat: add back indices

* feat: rework benchmarks

* docs: point to new benchmarks

+++ src/incremental/bucket_collection.rs
@@ -10,7 +10,7 @@ where
 {
     length: usize,
     haystacks: [&'a str; L],
-    idxs: [usize; L],
+    idxs: [u32; L],
     _phantom: PhantomData<N>,
 }
 
@@ -42,7 +42,7 @@ where
     pub fn add_haystack(
         &mut self,
         haystack: &'a str,
-        idx: usize,
+        idx: u32,
         buckets: &mut Vec<Box<dyn IncrementalBucketTrait>>,
     ) {
         self.haystacks[self.length] = haystack;
